+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * make sure path is not included more than once in filter list
 * smaller optimizations
 * updated User-Agent blacklist
+* fixed time on page test when grouping by page title
 
 ## 3.6.1